requirements.txt
Flask==2.0.2
gunicorn==19.9.0
line-bot-sdk==2.1.0
beautifulsoup4
requests